refactor(bitzel): Simplify build.rs to use TaskExecutor infrastructure
Refactored build.rs from manual implementation to use the existing
TaskExecutor infrastructure, reducing code from ~450 lines to ~250 lines.

Changes:
- Use TaskExecutor::new() instead of manual CAS/ActionCache setup
- Delegate orchestration to executor.execute_task() instead of manual execution
- Add expand_variables() helper for ${VAR} pattern expansion
- Use Pipeline's 3-stage process (discover → parse → build_graph)
- Create TaskSpec and let TaskExecutor handle:
  - Signature computation
  - Cache checking
  - Sandbox execution
  - Output collection
  - Cache storage

API fixes:
- Import BuildContext from layer_context (LayerBuildContext)
- Use recipe_extractor::BuildContext (different type) for ExtractionConfig
- Access Recipe from RecipeGraph via find_recipe() → get_recipe()
- Get task implementations from ParsedRecipe.task_impls instead of Recipe
- Use graph.statistics() for dependency count

The build command now successfully:
- Initializes TaskExecutor with content-addressable caching
- Discovers and parses 948 recipes in parallel
- Executes tasks with Linux namespace sandboxing
- Provides network isolation and resource limits
- Uses hash-based task signatures for hermetic builds

Preserved build_old.rs as reference for the manual implementation.
